Squashed 'lispBM/lispBM/' changes from 6199703d..43ebce71
43ebce71 Fix formatting int print.c and constants that default to float/double
fa220f0b update gc statistics collection a small bit
25b20686 fix get_heap_state
783d774e Arrays of 64bit values passes the associated tests. More tests are needed though
ff40d02f working on 64bit arrays. more testing needed
204bea39 CMP as a macro instead of 8 different functions
44bb152d added a set of math extensions grabbed from  Benjamin's bldc repo
9c34b4cd added string extensions from benjamins bldc repo
b4370155 merge in changes on master into dev64bit
c129344b fixed masking bug in fundamentals
ed3ab5be closed down ome warnings when building 64bit
d46564c9 expanded functionality of type-of to cover 64bit types and preparations for arrays of 64bit values
f7cb5538 added more 64bit functionality to fundamentals and a tiny amount of tests
4dafbe6c added decoders for 64bit values
63026a8a added 64bit value literals
8c1f0f0f made GC aware of lbm_memory allocated values
17148ada 64 bit values allocated on lbm_memory on 32 bit version
4445e5a8 sketch of encoders for 64 bit values
f636e64d suffix for the lisp size of int and uint are now i and u instead of i28 and u28
840723ca preparations and planning for 64bit types in both 32 and 64 bit versions.
13675dda 64bit up and running. but there are many TODOs
033bfd9a small fixes following bug hunt
f5c984de fix bug in representation of important masks and constants
d40b2437 Merge branch 'master' into dev64bit
06f9603f Merge branch 'master' into dev64bit
69950ba1 32 bit tests are ok
46b6fa28 Merge branch 'master' into dev64bit
b9e2c993 work in progress
8dff9b4d work in progress
a241aded work in progress
4738e0da update year in lbm_types
ca469923 made it possible to run the same tests as on the 32 bit version on the 64 bit code
85acec30 small tweaks tp create_ctx, use correct type
bf0286a7 Merge branch 'master' into dev64bit
824e1634 Merge branch 'master' into dev64bit
04a97f17 lots of testing needed
ea862cce a bunch of changes in preparation for 64bit compatibility.

git-subtree-dir: lispBM/lispBM
git-subtree-split: 43ebce71a281f4a158b8d2dec3fc537c26766ec3

+++ src/compression.c
@@ -193,10 +193,10 @@ int match_longest_key(char *string) {
 
   int longest_match_ix = -1;
   unsigned int longest_match_length = 0;
-  unsigned int n = strlen(string);
+  unsigned int n = (unsigned int)strlen(string);
 
   for (int i = 0; i < NUM_CODES; i ++) {
-    unsigned int s_len = strlen(codes[i][KEY]);
+    unsigned int s_len = (unsigned int)strlen(codes[i][KEY]);
     if (s_len <= n) {
       if (strncasecmp(codes[i][KEY], string, s_len) == 0) {
         if (s_len > longest_match_length) {
@@ -216,7 +216,7 @@ int match_longest_code(char *string, uint32_t start_bit, uint32_t total_bits) {
   unsigned int longest_match_length = 0;
 
   for (int i = 0; i < NUM_CODES; i++) {
-    unsigned int s_len = strlen(codes[i][CODE]);
+    unsigned int s_len = (unsigned int)strlen(codes[i][CODE]);
     if ((uint32_t)s_len <= bits_left) {
       bool match = true;
       for (uint32_t b = 0; b < (uint32_t)s_len; b ++) {
@@ -242,7 +242,7 @@ int match_longest_code(char *string, uint32_t start_bit, uint32_t total_bits) {
 int compressed_length(char *string) {
   uint32_t i = 0;
 
-  uint32_t n = strlen(string);
+  uint32_t n = (uint32_t)strlen(string);
   int comp_len = 0; // in bits
 
   bool string_mode = false;
@@ -292,9 +292,9 @@ int compressed_length(char *string) {
       }
 
       if (ix == -1)return -1;
-      unsigned int code_len = strlen(codes[ix][1]);
+      unsigned int code_len = (unsigned int)strlen(codes[ix][1]);
       comp_len += (int)code_len;
-      i += strlen(codes[ix][0]);
+      i += (unsigned int)strlen(codes[ix][0]);
     }
   }
   return comp_len;
@@ -324,7 +324,7 @@ void emit_string_char_code(char *compressed, char c, int *bit_pos) {
 }
 
 void emit_code(char *compressed, char *code, int *bit_pos) {
-  unsigned int n = strlen(code);
+  unsigned int n = (unsigned int)strlen(code);
 
   for (unsigned int i = 0; i < n; i ++) {
     int byte_ix = (*bit_pos) / 8;
@@ -388,7 +388,7 @@ char *lbm_compress(char *string, uint32_t *res_size) {
 
   bool string_mode = false;
   bool gobbling_whitespace = false;
-  uint32_t n = strlen(string);
+  uint32_t n = (uint32_t) strlen(string);
   uint32_t i = 0;
 
   while (i < n) {
@@ -442,7 +442,7 @@ char *lbm_compress(char *string, uint32_t *res_size) {
 
       emit_code(compressed, codes[ix][CODE], &bit_pos);
 
-      i += strlen(codes[ix][0]);
+      i += (unsigned int)strlen(codes[ix][0]);
     }
   }
 
@@ -488,7 +488,7 @@ int lbm_decompress_incremental(decomp_state *s, char *dest_buff, uint32_t dest_n
       s->last_string_char = 0;
     }
 
-    unsigned int n_bits_decoded = strlen(codes[ix][CODE]);
+    unsigned int n_bits_decoded = (unsigned int)strlen(codes[ix][CODE]);
     emit_key(dest_buff, codes[ix][KEY], (int)strlen(codes[ix][KEY]), &char_pos);
     s->i+=n_bits_decoded;
     return (int)char_pos;
